[fix] 알람 상태 dto =>  value 객체로 변환
