--- src/domain/user/dto/request/AlarmStatus.ts
@@ -1,10 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class AlarmStatus {
-  @IsNotEmpty()
-  content: string;
-
-  public getContent() {
-    return this.content;
-  }
-}
